Count method returns long instead of int

+++ src/main/java/WordCounter.java
@@ -17,7 +17,7 @@ public final class WordCounter {
 
     private final List<String> forbiddenWords = new ArrayList<>();
 
-    public int count(final String sentence) {
+    public long count(final String sentence) {
         if (isInvalidSentence(sentence)) {
             return 0;
         }
